[RENAME] Directory 오타 수정
